Allow chainloading EFI apps from loop mounts.
